make lawyer dasboard

+++ src/app/api/users/[id]/route.ts
@@ -1,4 +1,7 @@
 import { connect } from "@/dbConfig/dbConfig";
+import Consultations from "@/models/Consultations";
+import LawyerProfiles from "@/models/LawyerProfiles";
+import Reviews from "@/models/Reviews";
 import Users from "@/models/userModel";
 import { NextRequest, NextResponse } from "next/server";
 
@@ -63,7 +66,25 @@ export async function DELETE(
     await connect();
 
     const userId = params.params.id;
+    const user = await Users.findById(userId);
+    if(!user) {
+      return NextResponse.json({ message: "User not found" }, { status: 404 });
+    }
+    if (user.role === "Admin") {
+      return NextResponse.json({ message: "Cannot delete admin user" }, { status: 403 });
+    }
+    if (user.role === "Lawyer") {
+      const lawyerProfile = await LawyerProfiles.findOne({ user: userId });
+      if (lawyerProfile) {
+        await Consultations.deleteMany({ lawyer_id: lawyerProfile._id });
+        await Reviews.deleteMany({ lawyer_id: lawyerProfile._id });
+
+        await LawyerProfiles.findByIdAndDelete(lawyerProfile._id);
+      
+      }
+    }
     await Users.findByIdAndDelete(userId);
+    await Consultations.deleteMany({ client_id: userId });
 
     return NextResponse.json({ message: "Users deleted successfully" });
   } catch (error: any) {
